reafctor(bpf): move `icmp_wsum_accumulate` helper
This helper is shared by ICMPv6 and ICMPv4 and will be imported by both
in future refactors.
